requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
boto3==1.18.64
botocore==1.21.64
cached-property==1.5.2
certifi==2021.5.30
chardet==4.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.4
django-jalali-date==0.3.2
django-ratelimit==3.0.1
django-storages==1.12.2
gunicorn==20.1.0
idna==2.10
isodate==0.6.0
jdatetime==3.6.2
jmespath==0.10.0
lxml==4.6.3
Pillow==8.2.0
psycopg2-binary==2.9.1
pyotp==2.6.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.2.0
zeep==4.0.0
